requirements.txt
Flask==3.1.2
humanize==4.15.0
lxml==6.0.2
PyYAML==6.0.3
Requests==2.32.5
